Formulating GroupState class #41
This code was written with the game mechanics at mind. Heence, the concept of a game has been broadend to include rounds. The other major feature that this class will now deal with is the handling of players. It handles players as they join, wait and start games. If this code is to be implimented, the 'AI' players will need a mojor redesign to be able to interact with this code. I belive that this redesign will dictate how actual players will interact with eachother in the future.